client/New: Don't unlazy the gRPC connection when creating
When moving to gRPC 1.64 (commit 63b4688175) the usage of the deprecated
`grpc.DialContext` was replaced with `grpc.NewClient`. However, this
change also required to drop the `WithBlock` option, which made sure
that the connection is actually established before returning.

Now, `grpc.NewClient` doesn't attempt to perform the connection but
defers it to the actual first RPC.

Querying the default runtime on client creation breaks that property
depending on whether the default namespace is set or not.

This commit moves the `runtime` field initialization behind a
`sync.Once` and defers it to the first time the field actually needs to
be used.

Signed-off-by: Paweł Gronowski <pawel.gronowski@docker.com>